requirements.txt
openai>=1.14
pyyaml>=6.0
python-dotenv>=1.0
